# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Recipe Vault API
  description: Backend API for Recipe Vault mobile app
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.recipevault.com
    description: Production server

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /api/recipes:
    get:
      summary: Get all recipes
      tags:
        - Recipes
      responses:
        '200':
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a new recipe
      tags:
        - Recipes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/recipes/{recipe_id}:
    get:
      summary: Get recipe by ID
      tags:
        - Recipes
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: Update recipe
      tags:
        - Recipes
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdate'
      responses:
        '200':
          description: Recipe updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      summary: Delete recipe
      tags:
        - Recipes
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Recipe deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/recipes/search:
    get:
      summary: Search recipes
      tags:
        - Recipes
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/ocr:
    post:
      summary: Extract recipe from image using OCR
      tags:
        - OCR
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OCR result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OCRResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/parse-url:
    post:
      summary: Parse recipe from URL
      tags:
        - URL Parser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
      responses:
        '200':
          description: Parsed recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeCreate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Ingredient:
      type: object
      required:
        - id
        - recipe_id
        - name
        - order
      properties:
        id:
          type: string
          format: uuid
        recipe_id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          type: string
        unit:
          type: string
        order:
          type: integer

    IngredientCreate:
      type: object
      required:
        - name
        - order
      properties:
        name:
          type: string
        amount:
          type: string
        unit:
          type: string
        order:
          type: integer

    Step:
      type: object
      required:
        - id
        - recipe_id
        - description
        - order
      properties:
        id:
          type: string
          format: uuid
        recipe_id:
          type: string
          format: uuid
        description:
          type: string
        order:
          type: integer
        image_url:
          type: string
          format: uri

    StepCreate:
      type: object
      required:
        - description
        - order
      properties:
        description:
          type: string
        order:
          type: integer
        image_url:
          type: string
          format: uri

    Recipe:
      type: object
      required:
        - id
        - title
        - user_id
        - created_at
        - updated_at
        - ingredients
        - steps
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        image_url:
          type: string
          format: uri
        prep_time:
          type: integer
          description: Preparation time in minutes
        cook_time:
          type: integer
          description: Cooking time in minutes
        servings:
          type: integer
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        tags:
          type: array
          items:
            type: string

    RecipeCreate:
      type: object
      required:
        - title
        - ingredients
        - steps
      properties:
        title:
          type: string
        description:
          type: string
        image_url:
          type: string
          format: uri
        prep_time:
          type: integer
        cook_time:
          type: integer
        servings:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientCreate'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepCreate'
        tags:
          type: array
          items:
            type: string

    RecipeUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        image_url:
          type: string
          format: uri
        prep_time:
          type: integer
        cook_time:
          type: integer
        servings:
          type: integer
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientCreate'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepCreate'
        tags:
          type: array
          items:
            type: string

    OCRResult:
      type: object
      properties:
        text:
          type: string
          description: Extracted text from image
        recipe:
          $ref: '#/components/schemas/RecipeCreate'
          description: Parsed recipe data (if available)

    Error:
      type: object
      properties:
        detail:
          type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'